refactor(server): webhook 实现从 scheduler 迁入 server
webhook 是 server 的固有能力：model/store/执行链路/HTTP API 全部归
tiangong-server。execute_webhook 直接用 ServerCoreBackend 投递消息，
不经 scheduler 的执行抽象，不再依赖 SchedulerContext。

- scheduler：删除 webhook 模块与 RunTracker 枚举，execute_core 精简为
  只服务 job（Job 逻辑内联），execute_job_with_store 签名不变（plugin 透明）
- server：webhook/model.rs、store.rs 落地为完整实现（原为 re-export shim），
  新增 webhook/executor.rs 自包含执行（防并发→解析 session→记 WebhookRun
  →拼 [Webhook触发] 消息→经 ServerCoreBackend 投递→写回 session）
- api/webhook.rs：trigger/invoke 改调 crate::webhook::executor
- desktop：webhook CRUD 改读写 tiangong_server::webhook::WebhookStore，
  trigger 改经嵌入 server 的 HTTP 接口（需 server 在线）
- 测试：webhook store 测试随代码迁入 server crate

依赖方向无环：scheduler 只剩 job；webhook 全在 server；desktop 经 server
访问 webhook。job/cron、plugin-scheduler、REST 路由、HMAC 校验不变。
