+++ crates/tiangong-scheduler/src/lib.rs
@@ -3,7 +3,6 @@ use anyhow::{Result, bail};
 pub mod executor;
 pub mod model;
 pub mod store;
-pub mod webhook;
 
 pub use executor::SchedulerContext;
 
@@ -13,8 +12,7 @@ pub(crate) const fn default_true() -> bool {
 
 /// 把多行/空白折叠为单行，并要求结果非空。
 ///
-/// 定时任务与 webhook 的「任务名称 / 任务描述」均为展示为单行的字段，二者共用同一
-/// 归一化规则，避免卡片渲染出现跨行错位。
+/// 定时任务的「任务名称 / 任务描述」展示为单行字段，归一化避免卡片渲染跨行错位。
 pub(crate) fn normalize_required_single_line(field: &str, value: &str) -> Result<String> {
     let normalized = value
         .split(['\r', '\n'])
