--- crates/tiangong-scheduler/src/webhook/mod.rs
@@ -1,2 +1,0 @@
-pub mod model;
-pub mod store;
